Move SortOrder to global types
